[FEAT]: Reorganize plugins for better lazyloading handling
WIP :
    - TODO : Telescope extensions
