+++ lua/ls-devs/plugins/completion/cmp_npm.lua
@@ -1,0 +1,10 @@
+return {
+  "David-Kunz/cmp-npm",
+  dependencies = { "nvim-lua/plenary.nvim" },
+  ft = "json",
+  opts = {
+    ignore = {},
+    only_semantic_versions = false,
+    only_latest_version = false,
+  },
+}
