+++ lua/ls-devs/plugins/completion/cmp_nvim_lua.lua
@@ -1,0 +1,4 @@
+return {
+  "hrsh7th/cmp-nvim-lua",
+  ft = "lua",
+}
