+++ lua/ls-devs/plugins/completion/cmp_sass_variables.lua
@@ -1,0 +1,7 @@
+return {
+  "pontusk/cmp-sass-variables",
+  ft = { "css", "scss" },
+  config = function()
+    vim.g.sass_variables_file = "_variables.scss"
+  end,
+}
