+++ lua/ls-devs/plugins/completion/cmp_zsh.lua
@@ -1,0 +1,9 @@
+return {
+  "tamago324/cmp-zsh",
+  dependencies = { "Shougo/deol.nvim" },
+  ft = { "zsh", "deoledit" },
+  opts = {
+    zshrc = true,
+    filetypes = { "deoledit", "zsh" },
+  },
+}
