+++ lua/ls-devs/plugins/completion/crates.lua
@@ -1,0 +1,5 @@
+return {
+  "saecki/crates.nvim",
+  event = { "BufRead Cargo.toml" },
+  dependencies = { "nvim-lua/plenary.nvim" },
+}
