+++ lua/ls-devs/plugins/highlight/nvim_ts_autotag.lua
@@ -1,0 +1,6 @@
+return {
+  "windwp/nvim-ts-autotag",
+  opts = {},
+  dependencies = { "nvim-treesitter/nvim-treesitter" },
+  ft = { "typescript", "typescriptreact" },
+}
