+++ lua/ls-devs/plugins/highlight/treesitter_cpp.lua
@@ -1,0 +1,4 @@
+return {
+  "tree-sitter/tree-sitter-cpp",
+  ft = { "c", "cpp" },
+}
