+++ lua/ls-devs/plugins/lsp/config/mason_null_ls.lua
@@ -19,7 +19,6 @@ return function()
       "php-cs-fixer",
       "proselint",
       "hadolint",
-      "fourmolu",
       "clang-format",
       "cpplint",
       "bash-debug-adapter",
