+++ lua/ls-devs/plugins/lsp/config/mason_nvim_dap.lua
@@ -8,7 +8,6 @@ return function()
       "codelldb",
       "js",
       "chrome",
-      "haskell",
       "kotlin-debug-adapter",
     },
     automatic_setup = true,
