+++ lua/ls-devs/plugins/lsp/init.lua
@@ -49,11 +49,5 @@ return {
         },
       },
     },
-    -- Snippets
-    {
-      "L3MON4D3/LuaSnip",
-      version = "v2.*",
-      event = "InsertEnter",
-    },
   },
 }
