+++ lua/ls-devs/plugins/system/telescope.lua
@@ -76,11 +76,9 @@ return {
         },
       },
     })
+    -- TODO: REWORK EXTENSIONS TO LOAD THEM ONLY ON CMD
     telescope.load_extension("fzf")
     telescope.load_extension("media_files")
-    telescope.load_extension("aerial")
-    telescope.load_extension("neoclip")
-    telescope.load_extension("macros")
     telescope.load_extension("emoji")
     telescope.load_extension("noice")
     telescope.load_extension("luasnip")
@@ -197,11 +195,6 @@ return {
       "<cmd>Telescope emoji<CR>",
       desc = "Telescope Emojis",
     },
-    {
-      "<leader>fi",
-      "<cmd>Telescope import<CR>",
-      desc = "Telescope Import",
-    },
   },
   dependencies = {
     { "nvim-lua/plenary.nvim" },
@@ -213,83 +206,5 @@ return {
     { "benfowler/telescope-luasnip.nvim" },
     { "nvim-telescope/telescope-media-files.nvim" },
     { "xiyaowong/telescope-emoji.nvim" },
-    {
-      "AckslD/nvim-neoclip.lua",
-      dependencies = {
-        { "nvim-telescope/telescope.nvim" },
-        { "ibhagwan/fzf-lua" },
-        {
-          "kkharji/sqlite.lua",
-        },
-      },
-      opts = {
-        history = 1000,
-        enable_persistent_history = true,
-        length_limit = 1048576,
-        continuous_sync = true,
-        db_path = vim.fn.stdpath("data") .. "/databases/neoclip.sqlite3",
-        filter = nil,
-        preview = true,
-        prompt = nil,
-        default_register = '"',
-        default_register_macros = "q",
-        enable_macro_history = true,
-        content_spec_column = false,
-        disable_keycodes_parsing = false,
-        on_select = {
-          move_to_front = false,
-          close_telescope = true,
-        },
-        on_paste = {
-          set_reg = false,
-          move_to_front = false,
-          close_telescope = true,
-        },
-        on_replay = {
-          set_reg = false,
-          move_to_front = false,
-          close_telescope = true,
-        },
-        on_custom_action = {
-          close_telescope = true,
-        },
-        keys = {
-          telescope = {
-            i = {
-              select = "<cr>",
-              paste = "<A-p>",
-              paste_behind = "<A-k>",
-              replay = "<c-q>", -- replay a macro
-              delete = "<c-d>", -- delete an entry
-              edit = "<c-e>", -- edit an entry
-              custom = {},
-            },
-            n = {
-              select = "<cr>",
-              --- It is possible to map to more than one key.
-              paste = { "p", "<c-p>" },
-              paste_behind = "P",
-              replay = "q",
-              delete = "d",
-              edit = "e",
-              custom = {},
-            },
-          },
-          fzf = {
-            select = "default",
-            paste = "ctrl-p",
-            paste_behind = "ctrl-k",
-            custom = {},
-          },
-        },
-      },
-    },
-    {
-      "piersolenski/telescope-import.nvim",
-      dependencies = "nvim-telescope/telescope.nvim",
-      config = function()
-        require("telescope").load_extension("import")
-      end,
-    },
   },
 }
