+++ lua/ls-devs/plugins/ui/lualine.lua
@@ -102,7 +102,6 @@ return {
           no = colors.red,
           s = colors.orange,
           S = colors.orange,
-          [""] = colors.orange,
           ic = colors.yellow,
           R = colors.violet,
           Rv = colors.violet,
@@ -177,7 +176,6 @@ return {
           return msg
         end
         for _, client in ipairs(clients) do
-          ---@diagnostic disable-next-line: undefined-field
           local filetypes = client.config.filetypes
           if filetypes and vim.fn.index(filetypes, buf_ft) ~= -1 then
             return client.name
