+++ lua/ls-devs/plugins/utils/mini_comments.lua
@@ -2,7 +2,6 @@ return {
   "echasnovski/mini.comment",
   config = function()
     local comment = require("mini.comment")
-
     comment.setup({
       options = {
         custom_commentstring = function()
